requirements.txt
Flask
requests
python-dotenv
Pillow
numpy
opencv-python
pytesseract
gunicorn